One of these has to be a double if we want this modifier to work.

+++ src/main/java/com/gmail/nossr50/skills/smelting/FuelEfficiencyEventHandler.java
@@ -13,7 +13,7 @@ public class FuelEfficiencyEventHandler {
     }
 
     protected void calculateBurnModifier() {
-        this.burnModifier = 1 + ((manager.getSkillLevel() / Smelting.burnModifierMaxLevel) * Smelting.burnTimeMultiplier);
+        this.burnModifier = 1 + (((double) manager.getSkillLevel() / Smelting.burnModifierMaxLevel) * Smelting.burnTimeMultiplier);
     }
 
     protected void modifyBurnTime() {
